+++ packages/docs/content/not-in-sidebar/contact.mdx
@@ -3,10 +3,10 @@ title: Contact us
 intro: If you've got a question, idea, or suggestion, get in touch with the Design System team.
 ---
 
-import GithubIcon from '../../src/components/Icons/GithubIcon';
-import JiraIcon from '../../src/components/Icons/JiraIcon';
-import SlackIcon from '../../src/components/Icons/SlackIcon';
-import EmailIcon from '../../src/components/Icons/EmailIcon';
+import GithubIcon from '../../src/components/icons/GithubIcon';
+import JiraIcon from '../../src/components/icons/JiraIcon';
+import SlackIcon from '../../src/components/icons/SlackIcon';
+import EmailIcon from '../../src/components/icons/EmailIcon';
