--- packages/docs/src/components/Layout/index.ts
@@ -1,2 +1,0 @@
-export { default } from './Layout';
-export type { PageStatus } from './Layout';
