+++ packages/docs/src/components/content/ContentRenderer.tsx
