+++ packages/docs/src/components/content/ResponsiveExample.tsx
