+++ packages/docs/src/components/icons/EmailIcon.tsx
