+++ packages/docs/src/components/icons/SlackIcon.tsx
