+++ packages/docs/src/components/layout/PageHeader.tsx
@@ -4,7 +4,7 @@ import { withPrefix } from 'gatsby';
 import { Badge } from '@cmsgov/design-system';
 import { FrontmatterInterface } from '../../helpers/graphQLTypes';
 import { makeGithubUrl, makeSketchUrl, makeStorybookUrl } from '../../helpers/urlUtils';
-import GithubIcon from '../Icons/GithubIcon';
+import GithubIcon from '../icons/GithubIcon';
 
 type PageHeaderProps = {
   frontmatter?: FrontmatterInterface;
